requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-MySQLdb==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.1.1
WTForms==3.0.1
qrcode[pil]==7.4.2
Pillow==10.0.1
mysql-connector-python==8.1.0 